+++ cart-rest-api-for-woocommerce.php
@@ -5,7 +5,7 @@
  * Description: CoCart is a <strong>REST API for WooCommerce</strong>. It focuses on <strong>the front-end</strong> of the store to manage the shopping cart allowing developers to build a headless store.
  * Author:      Sébastien Dumont
  * Author URI:  https://sebastiendumont.com
- * Version:     2.6.0-beta.1
+ * Version:     2.6.0-beta.2
  * Text Domain: cart-rest-api-for-woocommerce
  * Domain Path: /languages/
  * Requires at least: 5.2
@@ -17,7 +17,6 @@
  */
 
 defined( 'ABSPATH' ) || exit;
-
 
 if ( ! defined( 'COCART_FILE' ) ) {
 	define( 'COCART_FILE', __FILE__ );
@@ -41,4 +40,15 @@ if ( ! function_exists( 'CoCart' ) ) {
 	}
 
 	CoCart();
+
+	/**
+	 * Load backend features only if COCART_WHITE_LABEL constant is
+	 * NOT set or IS set to false in user's wp-config.php file.
+	 */
+	if (
+		! defined( 'COCART_WHITE_LABEL' ) || false === COCART_WHITE_LABEL &&
+		is_admin() || ( defined( 'WP_CLI' ) && WP_CLI )
+	) {
+		include_once( untrailingslashit( plugin_dir_path( COCART_FILE ) ) . '/includes/admin/class-cocart-admin.php' );
+	}
 }