+++ includes/admin/views/html-notice-upgrade-warning.php
@@ -39,7 +39,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 		<div class="cocart-action">
 			<?php printf( '<a href="%1$s" class="button button-primary cocart-button" target="_blank">%2$s</a>', esc_url( 'https://cocart.xyz/contact/' ), esc_html__( 'Sign Up to Test', 'cart-rest-api-for-woocommerce' ) ); ?>
-			<span class="no-thanks"><a href="<?php echo esc_url( add_query_arg( 'hide_cocart_upgrade_notice', 'true' ) ); ?>" aria-label="<?php echo esc_html__( 'Hide this notice and ask me again another time', 'cart-rest-api-for-woocommerce' ); ?>"><?php echo esc_html__( 'Ask me another time', 'cart-rest-api-for-woocommerce' ); ?></a> / <a href="<?php echo esc_url( add_query_arg( 'hide_forever_cocart_upgrade_notice', 'true' ) ); ?>" aria-label="<?php echo esc_html__( 'Hide this notice forever.', 'cart-rest-api-for-woocommerce' ); ?>"><?php echo esc_html__( 'Don\'t ask me again', 'cart-rest-api-for-woocommerce' ); ?></a></span>
+			<span class="no-thanks"><a href="<?php echo esc_url( add_query_arg( 'hide_cocart_upgrade_notice', 'true', CoCart_Helpers::cocart_get_current_admin_url() ) ); ?>" aria-label="<?php echo esc_html__( 'Hide this notice and ask me again another time', 'cart-rest-api-for-woocommerce' ); ?>"><?php echo esc_html__( 'Ask me another time', 'cart-rest-api-for-woocommerce' ); ?></a> / <a href="<?php echo esc_url( add_query_arg( 'hide_forever_cocart_upgrade_notice', 'true', CoCart_Helpers::cocart_get_current_admin_url() ) ); ?>" aria-label="<?php echo esc_html__( 'Hide this notice forever.', 'cart-rest-api-for-woocommerce' ); ?>"><?php echo esc_html__( 'Don\'t ask me again', 'cart-rest-api-for-woocommerce' ); ?></a></span>
 		</div>
 	</div>
 </div>